requirements.txt
pandas==2.1.4
requests==2.31.0
SQLAlchemy==1.4.52
psycopg2-binary==2.9.9
apache-airflow==2.9.0
